requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
async-timeout==4.0.3
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
dataclasses-json==0.6.7
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.17
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-openai==1.2.1
langchain-protocol==0.0.14
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.8.0
marshmallow==3.26.2
multidict==6.7.1
mypy-extensions==1.1.0
numpy==2.2.6
openai==2.33.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
propcache==0.4.1
pydantic==2.13.3
pydantic-core==2.46.3
pydantic-settings==2.14.0
pypdf==6.10.2
python-dotenv==1.2.2
pyyaml==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.49
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
